RobotInfo public API migrated, tested
